requirements.txt
# Core Django (LTS - Updated for Feb 2026 Security Fixes)
Django==4.2.28
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.1

# Database
psycopg2-binary==2.9.9
django-environ==0.11.2

# Authentication & Security (DOWGRADED for Django 4.2 compatibility)
djangorestframework-simplejwt==5.3.1
django-allauth==0.58.2
django-guardian==2.4.0
django-two-factor-auth==1.15.1
cryptography==44.0.0
django-basicauth-0.5.3

# API Documentation
drf-spectacular==0.27.2

# File handling
django-storages==1.14.2
boto3==1.34.47
Pillow==10.3.0

# Email & Redis Cache
django-anymail==10.3
django-redis==5.4.0
hiredis==2.3.2

# Celery (DOWGRADED for Django 4.2 compatibility)
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
flower==2.0.1

# Monitoring & Logging
sentry-sdk==2.0.0
django-prometheus==2.3.1

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.1
python-magic==0.4.27
requests==2.32.5
setuptools==70.0.0

# Development Helpers
black==24.4.2
flake8==7.0.0
isort==5.13.2
pytest==9.0.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
Faker==25.0.0
django-debug-toolbar==4.3.0
django-extensions==3.2.3

# Production
gunicorn==22.0.0
whitenoise==6.6.0
django-health-check==3.18.2

# Security
django-otp==1.3.0
pyotp==2.9.0
qrcode==7.3.1

# Process and Utility
psutil==7.2.2

# GeoIP
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
frozenlist==1.8.0
geoip2==5.2.0
maxminddb==3.0.0
multidict==6.7.1
propcache==0.4.1
yarl==1.22.0

# Channels
channels==4.0.0
channels-redis==4.1.0

# Load Testing
locust==2.43.3
flask==3.1.2
flask-cors==6.0.2
flask-login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.7
greenlet==3.3.1
python-engineio==4.13.1
python-socketio==5.16.1
pywin32==311
pyzmq==27.1.0
simple-websocket==1.1.0
websocket-client==1.9.0
werkzeug==3.1.5
pytest<10,>=8.3.3
configargparse==1.7.1
bidict==0.23.1
blinker==1.9.0
wsproto==1.3.2
zope.event==6.1
zope.interface==8.2
brotli==1.2.0
pygments==2.19.2
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3